Remove redundant ibpconfig package in favor for single config package
